Move type converter infrastructure to different namespace.
