+++ src/FluentFiles.Tests/FixedLength/FixedLengthLineParserTests.cs
@@ -2,7 +2,7 @@ using Xunit;
 
 namespace FluentFiles.Tests.FixedLength
 {
-    using FluentFiles.Core.Base;
+    using FluentFiles.Core.Conversion;
     using FluentFiles.FixedLength;
     using FluentFiles.FixedLength.Implementation;
     using FluentFiles.Tests.Base.Entities;
